Submit libft

+++ libft/ft_strlen.c
@@ -6,15 +6,16 @@
 /*   By: younhwan <younhwan@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/04 23:55:31 by younhwan          #+#    #+#             */
-/*   Updated: 2022/07/04 23:59:38 by younhwan         ###   ########.fr       */
+/*   Updated: 2022/07/11 17:06:38 by younhwan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+// size_t
+size_t	ft_strlen(const char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len])
